requirements.txt
# ========================================
# VOLTIX AUDIT - Dépendances Python
# ========================================

# Framework mobile

Flask==3.1.3
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1

# Base de données
# sqlite3 est inclus avec Python (pas besoin d'installer)

# Génération de PDF
reportlab==4.4.10

# Traitement d'images et OCR (pour extraction facture)
Pillow==12.1.1
pytesseract==0.3.13

# Cryptographie (pour mots de passe)
bcrypt==5.0.0

# Envoi d'emails
python-emailer==1.5.4

# Gestion des dates
python-dateutil==2.9.0post0

# Validation des données
email-validator==2.3.0

# Graphiques pour les indicateurs de performance
matplotlib==3.10.8

# Manipulation de données
pandas==3.0.1

# Requests (pour API de paiement mobile si besoin)
requests==2.32.5
# OCR pour extraction de factures (Plan Entreprise)
pytesseract==0.3.13
Pillow==12.1.1

Werkzeug==3.1.6
gunicorn==25.1.0
